Move the api client from the daemon to mullvad-api
