+++ mullvad-api/src/lib.rs
@@ -25,6 +25,7 @@ use availability::ApiAvailability;
 pub mod rest;
 
 mod abortable_stream;
+pub mod api;
 mod https_client_with_sni;
 pub mod proxy;
 mod tls_stream;
